// src/types/core/asset/redeemable_nft/mint_moment_asset.ts
export type MintMomentProps = {
  nftId: string;
  data: string;
  dataMime: string;
  dataExtension: string;
  dataSize: number;
  dataProtocol: string;
  cover: string;
  coverMime: string;
  coverExtension: string;
  coverSize: number;
  coverProtocol: string;
  text: string;
};

export type MintMomentUI = MintMomentProps;